Refactor: Removed HelperFunctions.groovy, as they are now empty
