--- plugins/nf-co2footprint/src/main/nextflow/co2footprint/utils/HelperFunctions.groovy
@@ -1,44 +1,0 @@
-package nextflow.co2footprint.utils
-
-import nextflow.co2footprint.Logging.Markers
-
-import groovy.util.logging.Slf4j
-import nextflow.processor.TaskId
-import nextflow.trace.TraceRecord
-
-/**
- * Collection of helper functions.
- */
-@Slf4j
-class HelperFunctions {
-
-    /**
-     * Retrieves a value from the trace record by key, or returns a default if the value is missing.
-     * If the value is `null`, a warning is logged (once per unique dedupKey).
-     *
-     * @param trace        The TraceRecord containing task metrics.
-     * @param taskID       The ID of the current task (used for logging).
-     * @param key          The trace field key to retrieve (e.g. 'realtime', '%cpu', 'rss').
-     * @param defaultValue The fallback value to return if the key is missing or null.
-     * @param dedupKey     A unique identifier to deduplicate warning messages in the logs.
-     * @return             The value from the trace if present, otherwise the defaultValue.
-     */
-    static Object getTraceOrDefault(TraceRecord trace, TaskId taskID, String key, Object defaultValue, String dedupKey) {
-        def value = trace.get(key)
-        if (value == null) {
-            String warnMessage
-            if (key == '%cpu') {
-                def numCores = (defaultValue instanceof Number) ? defaultValue / 100 : defaultValue
-                warnMessage = "Missing trace value '${key}' for task ${taskID}, using default: 100% for all ${numCores} cores."
-            } else {
-                warnMessage = "Missing trace value '${key}' for task ${taskID}, using default: ${defaultValue}."
-            }
-            log.warn(
-                Markers.unique,
-                warnMessage,
-                dedupKey
-            )
-        }
-        return value ?: defaultValue
-    }
-}
